// src/app/modules/openapi/model/order-status-dto.ts
/**
 * E-Commerce Application
 *
 * Contact: sad@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OrderStatusDTO { 
    status: OrderStatusDTO.StatusEnum;
}
export namespace OrderStatusDTO {
    export const StatusEnum = {
        Processing: 'PROCESSING',
        Shipped: 'SHIPPED',
        Completed: 'COMPLETED',
        Canceled: 'CANCELED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}